Abstracted and simplified the code used to load and identify blogs which use the PQ
Now, all of that code can be properly tested WITHOUT having to bootstrap the CMS. The only things left in the CMS module are the call to the above method and the code extracting information from the blogs returned by the above method for the purposes of providing values for the template parameters (which is usually the only thing that SHOULD be in a CMS handler module).

Note that I also abstracted the test to see whether the blog has any async templates.  Any time there's complex code, look to abstract it out of its immediate surroundings and into its own method/class.